refactor: extract artist profile data controller from profile page
